Update version report
Update version report

+++ domain-core/src/main/java/pitayaa/nail/domain/report/detail/CustomerData.java
@@ -8,6 +8,6 @@ public class CustomerData {
 	String customerType;
 	Integer number;
 	Double rate;
-	Integer percentage;
+	Double percentage;
 	
 }
